Przeniesienie klasy node wewnątrz klasy StackList
